rename keyserver-test to mveloKeyServer-test
This way the name of the test matches the name of the module.
